Big file refactor - Change structure of fuzzy/crisp (#12) - Move common resource_activity_performance files under its package - Move specific CRISP/FUZZY files under their packages - Update test files
