+++ src/pix_framework/calendar/availability.py
@@ -7,7 +7,7 @@ from typing import List
 import pandas as pd
 import pytz
 
-from pix_framework.calendar.crisp_resource_calendar import RCalendar, Interval
+from pix_framework.discovery.resource_calendar_and_performance.crisp.resource_calendar import RCalendar, Interval
 
 
 def get_last_available_timestamp(start: pd.Timestamp, end: pd.Timestamp, schedule: RCalendar) -> pd.Timestamp:
