+++ src/pix_framework/discovery/resource_calendar_and_performance/crisp/discovery.py
@@ -2,8 +2,8 @@ from typing import List, Optional
 
 import pandas as pd
 
-from pix_framework.calendar.crisp_resource_calendar import RCalendar
-from pix_framework.discovery.calendar_discovery_parameters import CalendarDiscoveryParameters, CalendarType
+from pix_framework.discovery.resource_calendar_and_performance.crisp.resource_calendar import RCalendar
+from pix_framework.discovery.resource_calendar_and_performance.calendar_discovery_parameters import CalendarDiscoveryParameters, CalendarType
 from pix_framework.discovery.resource_calendar_and_performance.crisp.factory import CalendarFactory
 from pix_framework.discovery.resource_profiles import ResourceProfile
 from pix_framework.io.event_log import EventLogIDs
