+++ src/pix_framework/discovery/resource_calendar_and_performance/crisp/resource_activity_performance.py
@@ -3,8 +3,8 @@ from typing import List
 import pandas as pd
 
 from pix_framework.calendar.availability import absolute_unavailability_intervals_within
-from pix_framework.calendar.crisp_resource_calendar import RCalendar
-from pix_framework.discovery.resource_activity_performance import ActivityResourceDistribution, ResourceDistribution
+from pix_framework.discovery.resource_calendar_and_performance.crisp.resource_calendar import RCalendar
+from pix_framework.discovery.resource_calendar_and_performance.resource_activity_performance import ActivityResourceDistribution, ResourceDistribution
 from pix_framework.discovery.resource_profiles import ResourceProfile
 from pix_framework.io.event_log import EventLogIDs
 from pix_framework.statistics.distribution import get_best_fitting_distribution
