+++ src/pix_framework/discovery/resource_calendar_and_performance/resource_activity_performance.py
