+++ tests/pix_framework/discovery/resource_calendar_and_performance/crisp/test_factory.py
