+++ tests/pix_framework/discovery/resource_calendar_and_performance/crisp/test_resource_activity_performance.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import pytest
-from pix_framework.discovery.calendar_discovery_parameters import (
+from pix_framework.discovery.resource_calendar_and_performance.calendar_discovery_parameters import (
     CalendarDiscoveryParameters,
     CalendarType,
 )
