+++ tests/pix_framework/discovery/test_resource_model.py
@@ -2,9 +2,9 @@ from pathlib import Path
 
 import pandas as pd
 import pytest
-from pix_framework.calendar.crisp_resource_calendar import RCalendar
-from pix_framework.discovery.calendar_discovery_parameters import CalendarDiscoveryParameters, CalendarType
-from pix_framework.discovery.resource_activity_performance import ActivityResourceDistribution
+from pix_framework.discovery.resource_calendar_and_performance.crisp.resource_calendar import RCalendar
+from pix_framework.discovery.resource_calendar_and_performance.calendar_discovery_parameters import CalendarDiscoveryParameters, CalendarType
+from pix_framework.discovery.resource_calendar_and_performance.resource_activity_performance import ActivityResourceDistribution
 from pix_framework.discovery.resource_model import ResourceModel, discover_resource_model
 from pix_framework.discovery.resource_profiles import ResourceProfile
 from pix_framework.discovery.start_time_estimator.concurrency_oracle import OverlappingConcurrencyOracle
